Add email notification for user event and birthday
- Move comments into helpers to use anywheare
