+++ app/mailers/notification_mailer.rb
@@ -1,7 +1,10 @@
 class NotificationMailer < ApplicationMailer
-  def event_mail(user, events)
+  add_template_helper(PeopleHelper)
+
+  def event_mail(user, events, holidays)
     @user = user
     @events = events
+    @holidays = holidays
     mail to: @user.email, subject: "OYACOからのお知らせ"
   end
 end
