+++ test/fixtures/notification_mailer/event_mail.html
@@ -2,6 +2,11 @@
   <body>
     <h1>MyStringさん、<a href="https://oyaco.herokuapp.com">OYACO</a>からお知らせです。</h1>
 <p>
+<b>「お父さんの誕生日」（2016年2月22日）</b>まで、あと<b>83日</b>です。
+あと20年で、日本人女性の平均寿命（86.83）になります。
+<br>
+</p>
+<p>
 2015年9月29日は<b>「MyString」</b>です。
 
 <br>
